+++ apps/forum/src/app/(app)/users/[handle]/loading.tsx
@@ -1,0 +1,2 @@
+import AppLoading from "../../loading";
+export default AppLoading;